+++ ostap/core/run_ostap.py
@@ -278,7 +278,9 @@ else:
 
 # =============================================================================
 if arguments.Config :
-    cc  = os.environ.get ('OSTAP_CONFIG','').split( os.pathsep )
+    from ostap.utils.basic import get_env as ostap_getenv  
+    ## cc  = os.environ.get ('OSTAP_CONFIG','').split( os.pathsep )
+    cc  = ostap_getenv ('OSTAP_CONFIG','').split( os.pathsep )
     cc += arguments.Config
     cc  = os.pathsep.join ( cc )
     os.environ['OSTAP_CONFIG'] = cc